requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langgraph>=0.0.43
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
newspaper3k>=0.2.8
chromadb>=0.5.0
sentence-transformers>=2.7.0
scikit-learn>=1.5.0
numpy>=1.26.0
torch>=2.2.0; platform_system != 'Windows' or platform_machine != 'arm64'
torch>=2.2.0; platform_system == 'Windows' and platform_machine == 'AMD64'
transformers>=4.40.0
langchain-core>=0.1.45
tavily-python>=0.3.3
wikipedia>=1.4.0
orjson>=3.10.0
jinja2>=3.1.3
loguru>=0.7.2
pytest>=8.2.0
langdetect>=1.0.9
pandas>=2.1.0
ddgs>=1.0.0
pytest-asyncio>=0.23.0
langfuse>=2.0.0
